back-end/requirements.txt
flask==3.0.2
firebase-admin==6.2.0
python-dotenv==1.0.0
pyrebase4==4.5.0